Add files via upload
This is Console based notepad program using a linked list with 4 connections data structure. It's not complete because it has it's own limitations. Such as cursor movement, text editing and some edge cases. I'll be working on it from time to time. Feel free to work on it or make any suggestions.